--- TinyClicker/README.txt
@@ -1,29 +1,0 @@
-	
-	IMPORTANT:
-
-	Make sure to run only one instance of the LDPlayer. 
-	You can cover the emulator window with other windows, but do not minimize it to tray. You can move it elsewhere. 
-	Player interactions with the game are possible but not recommended. 
-
-	Do not hide the vertical panel on the right side of the LDPlayer with window controls (where the Home and Back buttons are located) 
-	as it is used by TinyClicker in cases when the restart of the game is needed. Restart is performed automatically when certain ads refuse to 
-	close or any other unexpected behaviour of the game occurs.
-
-
-	SETUP:
-
-     1. Prior to launching the clicker, set the LDPlayer resolution setting to Customize, with Width = 333, Height = 592, DPI = 150.
-	It is also recommended to set the "Fixed window size" to "Enable" at the "Other settings" tab, to prevent accidental resolution change.
-	This is important and TinyClicker won't work without the correct resolution.
-
-     2. If you don't have the VIP Tiny Tower package, open the file Config.txt, located inside the TinyClicker folder with notepad.
-	Locate the "VipPackage": setting and change it to "false" without quotes. If you have the VIP package leave the setting at true. 
-	You can also provide the current number of floors (e.g. 14 or 27 etc.) for automatic tower rebuilding and the elevator speed (e.g. 7.25 or 6 etc.).
-	Only use the format provided by examples in brackets.
-
-     3. When setup is done, simply launch LDPlayer, then the game and lastly the clicker. 
-	This setup is done once unless some of the parameters such as elevator speed or VIP status changes.
-
-     4. ???
-
-     5. Enjoy!
