+++ TinyClicker/src/Interfaces/ISampleImage.cs
@@ -1,0 +1,14 @@
+﻿using OpenCvSharp;
+using System;
+
+
+namespace TinyClicker.Interfaces;
+
+internal interface ISampleImage
+{
+    public int OrderedNumber { get; }
+    string ImageName { get; }
+    Mat Mat { get; }
+
+    Mat GenerateMat();
+}
