--- TinyClicker/src/Models/ISampleImage.cs
@@ -1,12 +1,0 @@
-﻿using OpenCvSharp;
-using System;
-
-
-namespace TinyClicker;
-
-internal interface ISampleImage
-{
-    Mat Mat { get; }
-
-    Mat GenerateMat(string imagePath);
-}
